feat: update HTML and XML services with enhanced validation, refactor namespace validator, and bump dependencies
