+++ packages/ace-linters/package.json
@@ -26,12 +26,12 @@
     "vite-plugin-node-polyfills": "^0.25.0",
     "rimraf": "^5.0.0",
     "vscode-html-languageservice": "^5.4.0",
-    "vscode-css-languageservice": "^6.3.5",
-    "vscode-json-languageservice": "^5.5.0",
+    "vscode-css-languageservice": "^6.3.10",
+    "vscode-json-languageservice": "^5.7.2",
     "luaparse": "^0.3.1",
     "showdown": "^2.1.0",
     "vscode-ws-jsonrpc": "^3.4.0",
-    "htmlhint": "^1.1.4",
+    "htmlhint": "^1.9.2",
     "@xml-tools/constraints": "^1.1.1",
     "@xml-tools/parser": "^1.0.11",
     "@xml-tools/ast": "^5.0.5",
